django_new_project/tmpl/_$project_name$_/requirements/common.txt
django==_$django_version$_
django-dotenv
ipython
